Module bevy_prototype_lyon::render

Render plugin

Structs

ShapeMaterial

A simple Material2d that renders with vertex colors.

ShapeMaterialPlugin

Plugin that provides a custom material for rendering [Shape]s

Constants

SHAPE_MATERIAL_SHADER_HANDLE

Handle to the custom shader with a unique random ID