Module bevy_prototype_lyon::prelude

Import this module as use bevy_prototype_lyon::prelude::* to get convenient imports.

Re-exports

pub use lyon_tessellation as tess;

pub use crate::draw::Fill;

pub use crate::draw::Stroke;

pub use crate::entity::Path;

pub use crate::entity::ShapeBundle;

pub use crate::geometry::Geometry;

pub use crate::geometry::GeometryBuilder;

pub use crate::path::PathBuilder;

pub use crate::path::ShapePath;

pub use crate::plugin::ShapePlugin;

pub use crate::shapes;

pub use crate::shapes::RectangleOrigin;

pub use crate::shapes::RegularPolygon;

pub use crate::shapes::RegularPolygonFeature;

Structs

FillOptions

Parameters for the fill tessellator.

StrokeOptions

Parameters for the tessellator.

Enums

FillRule

The fill rule defines how to determine what is inside and what is outside of the shape.

LineCap

Line cap as defined by the SVG specification.

LineJoin

Line join as defined by the SVG specification.

Orientation

Vertical or Horizontal.